yansi

Enum Color

pub enum Color {

    Primary,
    Fixed(u8),
    Rgb(u8, u8, u8),
    Black,
    Red,
    Green,
    Yellow,
    Blue,
    Magenta,
    Cyan,
    White,
    BrightBlack,
    BrightRed,
    BrightGreen,
    BrightYellow,
    BrightBlue,
    BrightMagenta,
    BrightCyan,
    BrightWhite,

}

Enum representing a terminal color.

Note: The color examples below are purely demonstrative. The actual color rendered depends entirely on the terminal and its configuration, the latter of which is entirely arbitrary.

Variants

Primary

Terminal primary color #9. (foreground code 39, background code 49).

This is the terminal’s defined “primary” color, that is, the configured default foreground and background colors. As such, this color as a foreground looks “good” against the terminal’s default background color, and this color is a “good” background color for the terminal’s default foreground color.

Fixed(u8)

A color from 0 to 255, for use in 256-color terminals.

Rgb(u8, u8, u8)

A 24-bit R G B “true color”, as specified by ISO-8613-3.

Black

Black #0 (foreground code 30, background code 40).

Red

Red #1 (foreground code 31, background code 41).

Green

Green: #2 (foreground code 32, background code 42).

Yellow

Yellow: #3 (foreground code 33, background code 43).

Blue

Blue: #4 (foreground code 34, background code 44).

Magenta

Magenta: #5 (foreground code 35, background code 45).

Cyan

Cyan: #6 (foreground code 36, background code 46).

White

White: #7 (foreground code 37, background code 47).

BrightBlack

Bright Black #0 (foreground code 90, background code 100).

BrightRed

Bright Red #1 (foreground code 91, background code 101).

BrightGreen

Bright Green: #2 (foreground code 92, background code 102).

BrightYellow

Bright Yellow: #3 (foreground code 93, background code 103).

BrightBlue

Bright Blue: #4 (foreground code 94, background code 104).

BrightMagenta

Bright Magenta: #5 (foreground code 95, background code 105).

BrightCyan

Bright Cyan: #6 (foreground code 96, background code 106).

BrightWhite

Bright White: #7 (foreground code 97, background code 107).

Implementations

impl Color

pub const fn foreground(self) -> Style

Returns a Style with a foreground color of self.

Example

use yansi::{Style, Color::*};

// A style with a foreground color of "yellow".
static DEBUG: Style = Yellow.foreground();

// This is equivalent to the above.
static DEBUG_S: Style = Style::new().fg(Yellow);

// The following two are equivalent. The latter is preferred.
static DEBUG_A: Style = Yellow.foreground().bold();
static DEBUG_B: Style = Yellow.bold();

pub const fn background(self) -> Style

Returns a Style with a background color of self.

Example

use yansi::{Style, Color::*};

// A style with a background color of "yellow".
static DEBUG: Style = Yellow.background();

// This is equivalent to the above.
static DEBUG_S: Style = Style::new().bg(Yellow);

// The following two are equivalent. The latter is preferred.
static DEBUG_A: Style = Yellow.background().green();
static DEBUG_B: Style = Green.on_yellow();

pub const fn bg(self, value: Color) -> Style

Returns a styled value derived from self with the background set to value.

This method should be used rarely. Instead, prefer to use color-specific builder methods like on_red() and on_green(), which have the same functionality but are pithier.

Example

Set background color to red using fg():

use yansi::{Paint, Color};

painted.bg(Color::Red);

Set background color to red using on_red().

use yansi::Paint;

painted.on_red();

pub const fn on_primary(self) -> Style

Returns self with the bg() set to [Color :: Primary].

Example

println!("{}", value.on_primary());

pub const fn on_fixed(self, color: u8) -> Style

Returns self with the bg() set to [Color :: Fixed].

Example

println!("{}", value.on_fixed(color));

pub const fn on_rgb(self, r: u8, g: u8, b: u8) -> Style

Returns self with the bg() set to [Color :: Rgb].

Example

println!("{}", value.on_rgb(r, g, b));

pub const fn on_black(self) -> Style

Returns self with the bg() set to [Color :: Black].

Example

println!("{}", value.on_black());

pub const fn on_red(self) -> Style

Returns self with the bg() set to [Color :: Red].

Example

println!("{}", value.on_red());

pub const fn on_green(self) -> Style

Returns self with the bg() set to [Color :: Green].

Example

println!("{}", value.on_green());

pub const fn on_yellow(self) -> Style

Returns self with the bg() set to [Color :: Yellow].

Example

println!("{}", value.on_yellow());

pub const fn on_blue(self) -> Style

Returns self with the bg() set to [Color :: Blue].

Example

println!("{}", value.on_blue());

pub const fn on_magenta(self) -> Style

Returns self with the bg() set to [Color :: Magenta].

Example

println!("{}", value.on_magenta());

pub const fn on_cyan(self) -> Style

Returns self with the bg() set to [Color :: Cyan].

Example

println!("{}", value.on_cyan());

pub const fn on_white(self) -> Style

Returns self with the bg() set to [Color :: White].

Example

println!("{}", value.on_white());

pub const fn on_bright_black(self) -> Style

Returns self with the bg() set to [Color :: BrightBlack].

Example

println!("{}", value.on_bright_black());

pub const fn on_bright_red(self) -> Style

Returns self with the bg() set to [Color :: BrightRed].

Example

println!("{}", value.on_bright_red());

pub const fn on_bright_green(self) -> Style

Returns self with the bg() set to [Color :: BrightGreen].

Example

println!("{}", value.on_bright_green());

pub const fn on_bright_yellow(self) -> Style

Returns self with the bg() set to [Color :: BrightYellow].

Example

println!("{}", value.on_bright_yellow());

pub const fn on_bright_blue(self) -> Style

Returns self with the bg() set to [Color :: BrightBlue].

Example

println!("{}", value.on_bright_blue());

pub const fn on_bright_magenta(self) -> Style

Returns self with the bg() set to [Color :: BrightMagenta].

Example

println!("{}", value.on_bright_magenta());

pub const fn on_bright_cyan(self) -> Style

Returns self with the bg() set to [Color :: BrightCyan].

Example

println!("{}", value.on_bright_cyan());

pub const fn on_bright_white(self) -> Style

Returns self with the bg() set to [Color :: BrightWhite].

Example

println!("{}", value.on_bright_white());

pub const fn attr(self, value: Attribute) -> Style

Enables the styling Attribute value.

This method should be used rarely. Instead, prefer to use attribute-specific builder methods like bold() and underline(), which have the same functionality but are pithier.

Example

Make text bold using attr():

use yansi::{Paint, Attribute};

painted.attr(Attribute::Bold);

Make text bold using using bold().

use yansi::Paint;

painted.bold();

pub const fn bold(self) -> Style

Returns self with the attr() set to [Attribute :: Bold].

Example

println!("{}", value.bold());

pub const fn dim(self) -> Style

Returns self with the attr() set to [Attribute :: Dim].

Example

println!("{}", value.dim());

pub const fn italic(self) -> Style

Returns self with the attr() set to [Attribute :: Italic].

Example

println!("{}", value.italic());

pub const fn underline(self) -> Style

Returns self with the attr() set to [Attribute :: Underline].

Example

println!("{}", value.underline());

pub const fn blink(self) -> Style

Returns self with the attr() set to [Attribute :: Blink].

Example

println!("{}", value.blink());

pub const fn rapid_blink(self) -> Style

Returns self with the attr() set to [Attribute :: RapidBlink].

Example

println!("{}", value.rapid_blink());

pub const fn invert(self) -> Style

Returns self with the attr() set to [Attribute :: Invert].

Example

println!("{}", value.invert());

pub const fn conceal(self) -> Style

Returns self with the attr() set to [Attribute :: Conceal].

Example

println!("{}", value.conceal());

pub const fn strike(self) -> Style

Returns self with the attr() set to [Attribute :: Strike].

Example

println!("{}", value.strike());

pub const fn quirk(self, value: Quirk) -> Style

Enables the yansi Quirk value.

This method should be used rarely. Instead, prefer to use quirk-specific builder methods like mask() and wrap(), which have the same functionality but are pithier.

Example

Enable wrapping using .quirk():

use yansi::{Paint, Quirk};

painted.quirk(Quirk::Wrap);

Enable wrapping using wrap().

use yansi::Paint;

painted.wrap();

pub const fn mask(self) -> Style

Returns self with the quirk() set to [Quirk :: Mask].

Example

println!("{}", value.mask());

pub const fn wrap(self) -> Style

Returns self with the quirk() set to [Quirk :: Wrap].

Example

println!("{}", value.wrap());

pub const fn linger(self) -> Style

Returns self with the quirk() set to [Quirk :: Linger].

Example

println!("{}", value.linger());

pub const fn clear(self) -> Style

👎Deprecated since 1.0.1: renamed to resetting() due to conflicts with Vec::clear(). The clear() method will be removed in a future release.

Returns self with the quirk() set to [Quirk :: Clear].

Example

println!("{}", value.clear());

pub const fn resetting(self) -> Style

Returns self with the quirk() set to [Quirk :: Resetting].

Example

println!("{}", value.resetting());

pub const fn bright(self) -> Style

Returns self with the quirk() set to [Quirk :: Bright].

Example

println!("{}", value.bright());

pub const fn on_bright(self) -> Style

Returns self with the quirk() set to [Quirk :: OnBright].

Example

println!("{}", value.on_bright());

pub const fn whenever(self, value: Condition) -> Style

Conditionally enable styling based on whether the Condition value applies. Replaces any previous condition.

See the crate level docs for more details.

Example

Enable styling painted only when both stdout and stderr are TTYs:

use yansi::{Paint, Condition};

painted.red().on_yellow().whenever(Condition::STDOUTERR_ARE_TTY);