yansi

Struct Condition

#[repr(transparent)]
pub struct Condition(pub fn() -> bool);

A function that decides whether styling should be applied.

A styling Condition can be specified globally via yansi::whenever() or locally to a specific style via the whenever() builder method. Any time a Painted value is formatted, both the local and global conditions are checked, and only when both evaluate to true is styling actually applied.

A Condition is nothing more than a function that returns a bool. The function is called each and every time a Painted is formatted, and so it is expected to be fast. All of the built-in conditions (except for their “live” variants) cache their first evaluation as a result: the Condition::cached() constructor can do the same for your conditions.

Built-In Conditions

yansi comes with built-in conditions for common scenarios that can be enabled via crate features:

feature(s)	condition	implication

detect-tty	TTY Detectors	std, is-terminal

detect-env	Environment Variable Checkers	std

[detect-tty, detect-env]	All Above, Combo Detectors	std, is-terminal

For example, to enable the TTY detectors, enable the detect-tty feature:

yansi = { version = "...", features = ["detect-tty"] }

To enable the TTY detectors, env-var checkers, and combo detectors, enable detect-tty and detect-env:

yansi = { version = "...", features = ["detect-tty", "detect-env"] }

use yansi::Condition;

yansi::whenever(Condition::TTY_AND_COLOR);

Custom Conditions

Custom, arbitrary conditions can be created with Condition::from() or Condition::cached().

use yansi::{Condition, Style, Color::*};

// Combine two conditions (`stderr` is a TTY, `CLICOLOR` is set) into one.
static STDERR_COLOR: Condition = Condition::from(||
    Condition::stderr_is_tty() && Condition::clicolor()
);

static DEBUG: Style = Yellow.bold().on_primary().invert().whenever(STDERR_COLOR);

Tuple Fields
0: fn() -> bool

The function that gets called to check the condition.

Implementations

impl Condition

pub const DEFAULT: Condition

A condition that evaluates to true if the OS supports coloring.

Uses Condition::os_support(). On Windows, this condition tries to enable coloring support on the first call and caches the result for subsequent calls. Outside of Windows, this always evaluates to true.

pub const ALWAYS: Condition

A condition that always evaluates to true.

pub const NEVER: Condition

A condition that always evaluated to false.

pub const fn from(f: fn() -> bool) -> Self

Creates a dynamically checked condition from a function f.

The function f is called anytime the condition is checked, including every time a style with the condition is used.

Example

use yansi::Condition;

fn some_function() -> bool {
    /* checking arbitrary conditions */
    todo!()
}

// Create a custom static condition from a function.
static MY_CONDITION: Condition = Condition::from(some_function);

// Create a condition on the stack from a function.
let my_condition = Condition::from(some_function);

// Create a static condition from a closure that becomes a `fn`.
static MY_CONDITION_2: Condition = Condition::from(|| false);

// Create a condition on the stack from a closure that becomes a `fn`.
let my_condition = Condition::from(|| some_function());

pub const fn cached(value: bool) -> Self

Creates a condition that is ALWAYS when value is true and NEVER otherwise.

Example

use yansi::Condition;

fn some_function() -> bool {
    /* checking arbitrary conditions */
    todo!()
}

// Cache the result of `some_function()` so it doesn't get called each
// time the condition needs to be checked.
let my_condition = Condition::cached(some_function());

pub const fn always() -> bool

The backing function for Condition::ALWAYS. Returns true always.

pub const fn never() -> bool

The backing function for Condition::NEVER. Returns false always.

pub fn os_support() -> bool

The backing function for Condition::DEFAULT.

Returns true if the current OS supports ANSI escape sequences for coloring. Outside of Windows, this always returns true. On Windows, the first call to this function attempts to enable support and returns whether it was successful every time thereafter.