Enum yansi::Attribute

pub enum Attribute {
    Bold,
    Dim,
    Italic,
    Underline,
    Blink,
    RapidBlink,
    Invert,
    Conceal,
    Strike,
}

Enum representing text attributes, largely for text formatting.

Text attributes are typically applied to a Style, Color, or Painted struct via the corresponding chainable builder methods such as bold() or italic(). The returned value will apply the attribute(s) when rendered or printed.

Attributes are idempotent, so applying an attribute more than once has no more affect than applying it once.

Terminal Support

Whether an applied attribute actually has an effect on how text is rendered in a terminal depends on the terminal’s support for the attribute as well as the terminal’s configuration. Common attributes, such as bold, dim, italic, underline, and strike typically have good support and are largely reliable. Less commonly supported attributes like conceal and invert will usually be supported by “modern” terminals. Rarely supprted attributes, like blink and rapid blink will usually have no effect when applied.

Example

use yansi::{Style, Color::Red};

/// A style with red foreground and every "common" attribute applied.
static MAD: Style = Red.bold().dim().italic().underline().strike();

Variants

Bold

Makes text bold.

Typically used via the bold() builder method.

Dim

Makes text dim.

Typically used via the dim() builder method.

Italic

Display text in italics.

Typically used via the italic() builder method.

Underline

Underline text.

Typically used via the underline() builder method.

Blink

Blink.

Typically used via the blink() builder method.

RapidBlink

Blink rapidly.

Typically used via the rapid_blink() builder method.

Invert

Invert (flip) the foreground and background colors.

Typically used via the invert() builder method.

Conceal

Conceal text.

Typically used via the conceal() builder method.

Strike

Display text with a strike through it.

Typically used via the strike() builder method.

Implementations

impl Attribute

pub const fn style(self) -> Style

Returns a Style with the attribute self enabled.

Example

use yansi::{Style, Attribute::Bold};

static EMBOLDEN: Style = Bold.style();