Struct yansi::Paint

pub struct Paint<T> { /* fields omitted */ }

A structure encapsulating all of the styling for a given item.

See the crate level documentation for usage information.

Methods

impl<T> Paint<T>

pub fn new(item: T) -> Paint<T>

Constructs a new Paint structure that encapsulates item. No styling is applied.

use yansi::Paint;

assert_eq!(Paint::new("hello!").to_string(), "hello!".to_string());

pub fn masked(item: T) -> Paint<T>

Constructs a new Paint structure that encapsulates item and masks it. No styling is applied. A masked item is not written out when painting is disabled during Display or Debug invocations. When painting is enabled, masking has no effect.

use yansi::Paint;

println!("{}Sprout!", Paint::masked("🌱 "));

pub fn black(item: T) -> Paint<T>

Constructs a new Paint structure that encapsulates item with the foreground color set to the name of this method.

use yansi::Paint;

println!("This is going to be blue: {}", Paint::blue("yay!"));

pub fn red(item: T) -> Paint<T>

Constructs a new Paint structure that encapsulates item with the foreground color set to the name of this method.

use yansi::Paint;

println!("This is going to be blue: {}", Paint::blue("yay!"));

pub fn green(item: T) -> Paint<T>

Constructs a new Paint structure that encapsulates item with the foreground color set to the name of this method.

use yansi::Paint;

println!("This is going to be blue: {}", Paint::blue("yay!"));

pub fn yellow(item: T) -> Paint<T>

Constructs a new Paint structure that encapsulates item with the foreground color set to the name of this method.

use yansi::Paint;

println!("This is going to be blue: {}", Paint::blue("yay!"));

pub fn blue(item: T) -> Paint<T>

Constructs a new Paint structure that encapsulates item with the foreground color set to the name of this method.

use yansi::Paint;

println!("This is going to be blue: {}", Paint::blue("yay!"));

pub fn purple(item: T) -> Paint<T>

Constructs a new Paint structure that encapsulates item with the foreground color set to the name of this method.

use yansi::Paint;

println!("This is going to be blue: {}", Paint::blue("yay!"));

pub fn cyan(item: T) -> Paint<T>

Constructs a new Paint structure that encapsulates item with the foreground color set to the name of this method.

use yansi::Paint;

println!("This is going to be blue: {}", Paint::blue("yay!"));

pub fn white(item: T) -> Paint<T>

Constructs a new Paint structure that encapsulates item with the foreground color set to the name of this method.

use yansi::Paint;

println!("This is going to be blue: {}", Paint::blue("yay!"));

pub fn rgb(r: u8, g: u8, b: u8, item: T) -> Paint<T>

Constructs a new Paint structure that encapsulates item with the foreground color set RGB color corresponding to r, g, b.

use yansi::Paint;

println!("This is going to be funky: {}", Paint::rgb(70, 130, 122, "hi!"));

pub fn fixed(color: u8, item: T) -> Paint<T>

Constructs a new Paint structure that encapsulates item with the foreground color set to the fixed color corresponding to color.

use yansi::Paint;

println!("This is going to be funky: {}", Paint::fixed(100, "hi!"));

pub fn fg(self, color: Color) -> Paint<T>

Sets the foreground to color.

use yansi::Paint;
use yansi::Color::Red;

println!("Red foreground: {}", Paint::new("hi!").fg(Red));

pub fn bg(self, color: Color) -> Paint<T>

Sets the background to color.

use yansi::Paint;
use yansi::Color::Yellow;

println!("Yellow background: {}", Paint::new("hi!").bg(Yellow));

pub fn mask(self) -> Paint<T>

Masks self. A masked item is not written out when painting is disabled during Display or Debug invocations. When painting is enabled, masking has no effect.

use yansi::Paint;

println!("{}Something happened.", Paint::red("Whoops! ").mask());

pub fn bold(self) -> Paint<T>

Enables the styling corresponding to the name of this method.

use yansi::Paint;

println!("Red, underlined: {}", Paint::red("beep.").underline());

pub fn dimmed(self) -> Paint<T>

Enables the styling corresponding to the name of this method.

use yansi::Paint;

println!("Red, underlined: {}", Paint::red("beep.").underline());

pub fn italic(self) -> Paint<T>

Enables the styling corresponding to the name of this method.

use yansi::Paint;

println!("Red, underlined: {}", Paint::red("beep.").underline());

pub fn underline(self) -> Paint<T>

Enables the styling corresponding to the name of this method.

use yansi::Paint;

println!("Red, underlined: {}", Paint::red("beep.").underline());

pub fn blink(self) -> Paint<T>

Enables the styling corresponding to the name of this method.

use yansi::Paint;

println!("Red, underlined: {}", Paint::red("beep.").underline());

pub fn invert(self) -> Paint<T>

Enables the styling corresponding to the name of this method.

use yansi::Paint;

println!("Red, underlined: {}", Paint::red("beep.").underline());

pub fn hidden(self) -> Paint<T>

Enables the styling corresponding to the name of this method.

use yansi::Paint;

println!("Red, underlined: {}", Paint::red("beep.").underline());

pub fn strikethrough(self) -> Paint<T>

Enables the styling corresponding to the name of this method.

use yansi::Paint;

println!("Red, underlined: {}", Paint::red("beep.").underline());

impl Paint<()>

pub fn disable()

Disables coloring globally.

This method is only available when the "nightly" feature is enabled.

Example

use yansi::Paint;

// With coloring enabled, ANSI color codes are emitted.
assert_ne!(Paint::green("go").to_string(), "go".to_string());

// With coloring disabled, ANSI color codes are _not_ emitted.
Paint::disable();
assert_eq!(Paint::green("go").to_string(), "go".to_string());

pub fn enable()

Enables coloring globally. Coloring is enabled by default, so this method should only be called to re enable coloring.

This method is only available when the "nightly" feature is enabled.

Example

use yansi::Paint;

// With coloring disabled, ANSI color codes are _not_ emitted.
Paint::disable();
assert_eq!(Paint::green("go").to_string(), "go".to_string());

// Reenabling causes color code to be emitted.
Paint::enable();
assert_ne!(Paint::green("go").to_string(), "go".to_string());

pub fn enable_windows_ascii() -> bool

Enables ASCII terminal escape sequences on Windows consoles when possible. Returns true if escape sequence support was successfully enabled and false otherwise. On non-Windows targets, this method always returns true.

Support for escape sequences in Windows consoles was added in the Windows 10 anniversary update. For targets with older Windows installations, this method is expected to return false.

Example

use yansi::Paint;

// A best-effort Windows ASCII terminal support enabling.
Paint::enable_windows_ascii();