yansi

Struct Style

pub struct Style {
    pub foreground: Option<Color>,
    pub background: Option<Color>,
    pub condition: Option<Condition>,
    /* private fields */
}

A set of styling options.

Equivalence and Ordering

Only a style’s foreground, background, and set of attributes are considered when testing for equivalence or producing an ordering via PartialEq or Eq, and PartialOrd or Ord. A style’s quirks and conditions are ignored.

Fields
foreground: Option<Color>

The foreground color. Defaults to None.

use yansi::{Style, Color};

assert_eq!(Style::new().foreground, None);
assert_eq!(Style::new().green().foreground, Some(Color::Green));

background: Option<Color>

The background color. Defaults to None.

use yansi::{Style, Color};

assert_eq!(Style::new().background, None);
assert_eq!(Style::new().on_red().background, Some(Color::Red));

condition: Option<Condition>

The condition.

To check a style’s condition directly, use Style::enabled():

use yansi::{Style, Condition};

let style = Style::new().whenever(Condition::ALWAYS);
assert!(style.enabled());

let style = Style::new().whenever(Condition::NEVER);
assert!(!style.enabled());

Implementations

impl Style

pub const fn new() -> Style

Returns a new style with no foreground or background, no attributes or quirks, and Condition::DEFAULT.

This is the default returned by Default::default().

Example

use yansi::Style;

assert_eq!(Style::new(), Style::default());

pub fn enabled(&self) -> bool

Returns true if this style is enabled, based on condition.

Note: For a style to be effected, both this method and yansi::is_enabled() must return true.

When there is no condition set, this method always returns true. When a condition has been set, this evaluates the condition and returns the result.

Example

use yansi::{Style, Condition};

let style = Style::new().whenever(Condition::ALWAYS);
assert!(style.enabled());

let style = Style::new().whenever(Condition::NEVER);
assert!(!style.enabled());

pub fn fmt_prefix(&self, f: &mut dyn Write) -> Result

Writes the ANSI code prefix for the currently set styles.

This method is intended to be used inside of fmt::Display and fmt::Debug implementations for custom or specialized use-cases. Most users should use Painted for all painting needs.

This method writes the ANSI code prefix irrespective of whether painting is currently enabled or disabled. To write the prefix only if painting is enabled, condition a call to this method on is_enabled().

Example

use core::fmt;
use yansi::Style;

struct CustomItem {
    item: u32,
    style: Style
}

impl fmt::Display for CustomItem {
    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
        self.style.fmt_prefix(f)?;
        write!(f, "number: {}", self.item)?;
        self.style.fmt_suffix(f)
    }
}

pub fn prefix(&self) -> Cow<'static, str>

Returns the ANSI code sequence prefix for the style as a string.

This returns a string with the exact same sequence written by fmt_prefix(). See that method for details.

pub fn fmt_suffix(&self, f: &mut dyn Write) -> Result

Writes the ANSI code sequence suffix for the style.

This method is intended to be used inside of fmt::Display and fmt::Debug implementations for custom or specialized use-cases. Most users should use Painted for all painting needs.

This method writes the ANSI code suffix irrespective of whether painting is currently enabled or disabled. To write the suffix only if painting is enabled, condition a call to this method on is_enabled().

Example

use core::fmt;
use yansi::Style;

struct CustomItem {
    item: u32,
    style: Style
}

impl fmt::Display for CustomItem {
    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
        self.style.fmt_prefix(f)?;
        write!(f, "number: {}", self.item)?;
        self.style.fmt_suffix(f)
    }
}

pub fn suffix(&self) -> Cow<'static, str>

Returns the ANSI code sequence suffix for the style as a string.

This returns a string with the exact same sequence written by fmt_suffix(). See that method for details.

pub const fn fg(self, value: Color) -> Self

Returns a styled value derived from self with the foreground set to value.

This method should be used rarely. Instead, prefer to use color-specific builder methods like red() and green(), which have the same functionality but are pithier.

Example

Set foreground color to white using fg():

use yansi::{Paint, Color};

painted.fg(Color::White);

Set foreground color to white using white().

use yansi::Paint;

painted.white();

pub const fn primary(self) -> Self

Returns self with the fg() set to [Color :: Primary].

Example

println!("{}", value.primary());

pub const fn fixed(self, color: u8) -> Self

Returns self with the fg() set to [Color :: Fixed].

Example

println!("{}", value.fixed(color));

pub const fn rgb(self, r: u8, g: u8, b: u8) -> Self

Returns self with the fg() set to [Color :: Rgb].

Example

println!("{}", value.rgb(r, g, b));

pub const fn black(self) -> Self

Returns self with the fg() set to [Color :: Black].

Example

println!("{}", value.black());

pub const fn red(self) -> Self

Returns self with the fg() set to [Color :: Red].

Example

println!("{}", value.red());

pub const fn green(self) -> Self

Returns self with the fg() set to [Color :: Green].

Example

println!("{}", value.green());

pub const fn yellow(self) -> Self

Returns self with the fg() set to [Color :: Yellow].

Example

println!("{}", value.yellow());

pub const fn blue(self) -> Self

Returns self with the fg() set to [Color :: Blue].

Example

println!("{}", value.blue());

pub const fn magenta(self) -> Self

Returns self with the fg() set to [Color :: Magenta].

Example

println!("{}", value.magenta());

pub const fn cyan(self) -> Self

Returns self with the fg() set to [Color :: Cyan].

Example

println!("{}", value.cyan());

pub const fn white(self) -> Self

Returns self with the fg() set to [Color :: White].

Example

println!("{}", value.white());

pub const fn bright_black(self) -> Self

Returns self with the fg() set to [Color :: BrightBlack].

Example

println!("{}", value.bright_black());

pub const fn bright_red(self) -> Self

Returns self with the fg() set to [Color :: BrightRed].

Example

println!("{}", value.bright_red());

pub const fn bright_green(self) -> Self

Returns self with the fg() set to [Color :: BrightGreen].

Example

println!("{}", value.bright_green());

pub const fn bright_yellow(self) -> Self

Returns self with the fg() set to [Color :: BrightYellow].

Example

println!("{}", value.bright_yellow());

pub const fn bright_blue(self) -> Self

Returns self with the fg() set to [Color :: BrightBlue].

Example

println!("{}", value.bright_blue());

pub const fn bright_magenta(self) -> Self

Returns self with the fg() set to [Color :: BrightMagenta].

Example

println!("{}", value.bright_magenta());

pub const fn bright_cyan(self) -> Self

Returns self with the fg() set to [Color :: BrightCyan].

Example

println!("{}", value.bright_cyan());

pub const fn bright_white(self) -> Self

Returns self with the fg() set to [Color :: BrightWhite].

Example

println!("{}", value.bright_white());

pub const fn bg(self, value: Color) -> Self

Returns a styled value derived from self with the background set to value.

This method should be used rarely. Instead, prefer to use color-specific builder methods like on_red() and on_green(), which have the same functionality but are pithier.

Example

Set background color to red using fg():

use yansi::{Paint, Color};

painted.bg(Color::Red);

Set background color to red using on_red().

use yansi::Paint;

painted.on_red();

pub const fn on_primary(self) -> Self

Returns self with the bg() set to [Color :: Primary].

Example

println!("{}", value.on_primary());

pub const fn on_fixed(self, color: u8) -> Self

Returns self with the bg() set to [Color :: Fixed].

Example

println!("{}", value.on_fixed(color));

pub const fn on_rgb(self, r: u8, g: u8, b: u8) -> Self

Returns self with the bg() set to [Color :: Rgb].

Example

println!("{}", value.on_rgb(r, g, b));

pub const fn on_black(self) -> Self

Returns self with the bg() set to [Color :: Black].

Example

println!("{}", value.on_black());

pub const fn on_red(self) -> Self

Returns self with the bg() set to [Color :: Red].

Example

println!("{}", value.on_red());

pub const fn on_green(self) -> Self

Returns self with the bg() set to [Color :: Green].

Example

println!("{}", value.on_green());

pub const fn on_yellow(self) -> Self

Returns self with the bg() set to [Color :: Yellow].

Example

println!("{}", value.on_yellow());

pub const fn on_blue(self) -> Self

Returns self with the bg() set to [Color :: Blue].

Example

println!("{}", value.on_blue());

pub const fn on_magenta(self) -> Self

Returns self with the bg() set to [Color :: Magenta].

Example

println!("{}", value.on_magenta());

pub const fn on_cyan(self) -> Self

Returns self with the bg() set to [Color :: Cyan].

Example

println!("{}", value.on_cyan());

pub const fn on_white(self) -> Self

Returns self with the bg() set to [Color :: White].

Example

println!("{}", value.on_white());

pub const fn on_bright_black(self) -> Self

Returns self with the bg() set to [Color :: BrightBlack].

Example

println!("{}", value.on_bright_black());

pub const fn on_bright_red(self) -> Self

Returns self with the bg() set to [Color :: BrightRed].

Example

println!("{}", value.on_bright_red());

pub const fn on_bright_green(self) -> Self

Returns self with the bg() set to [Color :: BrightGreen].

Example

println!("{}", value.on_bright_green());

pub const fn on_bright_yellow(self) -> Self

Returns self with the bg() set to [Color :: BrightYellow].

Example

println!("{}", value.on_bright_yellow());

pub const fn on_bright_blue(self) -> Self

Returns self with the bg() set to [Color :: BrightBlue].

Example

println!("{}", value.on_bright_blue());

pub const fn on_bright_magenta(self) -> Self

Returns self with the bg() set to [Color :: BrightMagenta].

Example

println!("{}", value.on_bright_magenta());

pub const fn on_bright_cyan(self) -> Self

Returns self with the bg() set to [Color :: BrightCyan].

Example

println!("{}", value.on_bright_cyan());

pub const fn on_bright_white(self) -> Self

Returns self with the bg() set to [Color :: BrightWhite].

Example

println!("{}", value.on_bright_white());

pub const fn attr(self, value: Attribute) -> Self

Enables the styling Attribute value.

This method should be used rarely. Instead, prefer to use attribute-specific builder methods like bold() and underline(), which have the same functionality but are pithier.

Example

Make text bold using attr():

use yansi::{Paint, Attribute};

painted.attr(Attribute::Bold);

Make text bold using using bold().

use yansi::Paint;

painted.bold();

pub const fn bold(self) -> Self

Returns self with the attr() set to [Attribute :: Bold].

Example

println!("{}", value.bold());

pub const fn dim(self) -> Self

Returns self with the attr() set to [Attribute :: Dim].

Example

println!("{}", value.dim());

pub const fn italic(self) -> Self

Returns self with the attr() set to [Attribute :: Italic].

Example

println!("{}", value.italic());

pub const fn underline(self) -> Self

Returns self with the attr() set to [Attribute :: Underline].

Example

println!("{}", value.underline());

pub const fn blink(self) -> Self

Returns self with the attr() set to [Attribute :: Blink].

Example

println!("{}", value.blink());

pub const fn rapid_blink(self) -> Self

Returns self with the attr() set to [Attribute :: RapidBlink].

Example

println!("{}", value.rapid_blink());

pub const fn invert(self) -> Self

Returns self with the attr() set to [Attribute :: Invert].

Example

println!("{}", value.invert());

pub const fn conceal(self) -> Self

Returns self with the attr() set to [Attribute :: Conceal].

Example

println!("{}", value.conceal());

pub const fn strike(self) -> Self

Returns self with the attr() set to [Attribute :: Strike].

Example

println!("{}", value.strike());

pub const fn quirk(self, value: Quirk) -> Self

Enables the yansi Quirk value.

This method should be used rarely. Instead, prefer to use quirk-specific builder methods like mask() and wrap(), which have the same functionality but are pithier.

Example

Enable wrapping using .quirk():

use yansi::{Paint, Quirk};

painted.quirk(Quirk::Wrap);

Enable wrapping using wrap().

use yansi::Paint;

painted.wrap();

pub const fn mask(self) -> Self

Returns self with the quirk() set to [Quirk :: Mask].

Example

println!("{}", value.mask());

pub const fn wrap(self) -> Self

Returns self with the quirk() set to [Quirk :: Wrap].

Example

println!("{}", value.wrap());

pub const fn linger(self) -> Self

Returns self with the quirk() set to [Quirk :: Linger].

Example

println!("{}", value.linger());

pub const fn clear(self) -> Self

👎Deprecated since 1.0.1: renamed to resetting() due to conflicts with Vec::clear(). The clear() method will be removed in a future release.

Returns self with the quirk() set to [Quirk :: Clear].

Example

println!("{}", value.clear());

pub const fn resetting(self) -> Self

Returns self with the quirk() set to [Quirk :: Resetting].

Example

println!("{}", value.resetting());

pub const fn bright(self) -> Self

Returns self with the quirk() set to [Quirk :: Bright].

Example

println!("{}", value.bright());

pub const fn on_bright(self) -> Self

Returns self with the quirk() set to [Quirk :: OnBright].

Example

println!("{}", value.on_bright());

pub const fn whenever(self, value: Condition) -> Self

Conditionally enable styling based on whether the Condition value applies. Replaces any previous condition.

See the crate level docs for more details.

Example

Enable styling painted only when both stdout and stderr are TTYs:

use yansi::{Paint, Condition};

painted.red().on_yellow().whenever(Condition::STDOUTERR_ARE_TTY);